PLT-1392 Moved to new package
