+++ examples/poaps/backend/src/index.ts
@@ -1,21 +1,10 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
+import { measurePerformance } from '@poap-xyz/performance';
 import { PoapsClient } from '@poap-xyz/poaps';
 import { PoapCompass } from '@poap-xyz/providers';
 import { fetch_multiple_poaps } from './methods/fetch_multiple_poaps';
 import { fetch_single_poap } from './methods/fetch_single_poap';
 import { fetch_multiple_poaps_by_collector } from './methods/fetch_multiple_poaps_by_collector';
 import { fetch_multiple_poaps_by_drop_id } from './methods/fetch_multiple_poaps_by_drop_id';
-import { performance } from 'perf_hooks';
-
-async function measurePerformance(
-  fn: () => Promise<void>,
-  name: string,
-): Promise<void> {
-  const start = performance.now();
-  await fn();
-  const end = performance.now();
-  console.log(`${name} took ${end - start} milliseconds to execute.`);
-}
 
 async function main(): Promise<void> {
   // Use your library here
