+++ packages/performance/rollup.config.js
@@ -1,0 +1,3 @@
+import configs from '../../rollup.base.config';
+
+export default configs;
